chore: move stuff to constants
